The lags are what's leaking data
As it turns out, the lag features are basically leaking the temperature. We're also trying to predict more than one hour which I realize we may not be doing at all. This may be useless right now because it's not being forced to predict in advance without the features. I'll come back to this later